+++ common/usbx_host_controllers/src/ux_hcd_ehci_asynchronous_endpoint_create.c
@@ -35,7 +35,7 @@
 /*  FUNCTION                                               RELEASE        */ 
 /*                                                                        */ 
 /*    _ux_hcd_ehci_asynchronous_endpoint_create           PORTABLE C      */ 
-/*                                                           6.1.2        */
+/*                                                           6.x          */
 /*  AUTHOR                                                                */
 /*                                                                        */
 /*    Chaoqiong Xiao, Microsoft Corporation                               */
@@ -75,6 +75,9 @@
 /*  11-09-2020     Chaoqiong Xiao           Modified comment(s),          */
 /*                                            fixed compile warnings,     */
 /*                                            resulting in version 6.1.2  */
+/*  xx-xx-xxxx     Chaoqiong Xiao           Modified comment(s),          */
+/*                                            fixed compile warnings,     */
+/*                                            resulting in version 6.x    */
 /*                                                                        */
 /**************************************************************************/
 UINT  _ux_hcd_ehci_asynchronous_endpoint_create(UX_HCD_EHCI *hcd_ehci, UX_ENDPOINT *endpoint)
@@ -102,7 +105,7 @@ UX_EHCI_LINK_POINTER    queue_head;
     ed -> REF_AS.INTR.ux_ehci_ed_endpoint =  endpoint;
 
     /* Set the default MPS Capability info in the ED.  */
-    ed -> ux_ehci_ed_cap0 =  endpoint -> ux_endpoint_descriptor.wMaxPacketSize << UX_EHCI_QH_MPS_LOC;
+    ed -> ux_ehci_ed_cap0 =  (ULONG)endpoint -> ux_endpoint_descriptor.wMaxPacketSize << UX_EHCI_QH_MPS_LOC;
     
     /* Set the default NAK reload count.  */
     ed -> ux_ehci_ed_cap0 |=  UX_EHCI_QH_NCR;
